+++ src/main/java/net/sashiro/additionalvanillastuff/data/DataGen.java
@@ -1,10 +1,10 @@
 package net.sashiro.additionalvanillastuff.data;
-/*
+
 import net.minecraft.data.DataGenerator;
 import net.minecraftforge.common.data.ExistingFileHelper;
+import net.minecraftforge.data.event.GatherDataEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.forge.event.lifecycle.GatherDataEvent;
 import net.sashiro.additionalvanillastuff.AdditionalVanillaStuff;
 import net.sashiro.additionalvanillastuff.data.generators.*;
 
@@ -15,12 +15,11 @@ public class DataGen {
         DataGenerator gen = e.getGenerator();
         ExistingFileHelper existingFileHelper = e.getExistingFileHelper();
 
-        gen.addProvider(new GenBlockStateProvider(gen, existingFileHelper));
-        gen.addProvider(new GenItemModelProvider(gen, existingFileHelper));
-        gen.addProvider(new GenLanguageProvider(gen, "en_us"));
-        gen.addProvider(new GenRecipeProvider(gen));
-        gen.addProvider(new GenLootTableProvider(gen));
-        gen.addProvider(new GenTagProvider(gen, AdditionalVanillaStuff.MOD_ID, existingFileHelper));
+        gen.addProvider(true, new GenBlockStateProvider(gen, existingFileHelper));
+        gen.addProvider(true, new GenItemModelProvider(gen, existingFileHelper));
+        gen.addProvider(true, new GenLanguageProvider(gen, "en_us"));
+        gen.addProvider(true, new GenRecipeProvider(gen));
+        gen.addProvider(true, new GenLootTableProvider(gen));
+        gen.addProvider(true, new GenTagProvider(gen, AdditionalVanillaStuff.MOD_ID, existingFileHelper));
     }
 }
- */