+++ src/main/java/net/sashiro/additionalvanillastuff/data/generators/GenRecipeProvider.java
@@ -1,12 +1,13 @@
 package net.sashiro.additionalvanillastuff.data.generators;
-/*
+
 import net.minecraft.data.DataGenerator;
 import net.minecraft.data.recipes.FinishedRecipe;
 import net.minecraft.data.recipes.RecipeProvider;
 import net.minecraft.data.recipes.ShapedRecipeBuilder;
+import net.minecraft.data.recipes.SingleItemRecipeBuilder;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.block.Blocks;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
 import net.sashiro.additionalvanillastuff.AdditionalVanillaStuff;
@@ -25,17 +26,19 @@ public class GenRecipeProvider extends RecipeProvider {
     protected void buildCraftingRecipes(Consumer<FinishedRecipe> consumer) {
 
         for (RegistryObject<Block> block : ModRegistryEvent.BLOCK_REGISTRY.getEntries()) {
-            String loc = block.get().getRegistryName().getPath();
-            Block vanillaBLock = Blocks.AIR;
+            assert false;
+            String loc = block.get().getDescriptionId().replace("block.additionalvanillastuff.", "");
+            Block vanillaBLock;
             for (Block mcBlock : ForgeRegistries.BLOCKS) {
-                String blockName = block.get().getRegistryName().getPath().replace("_wall", "").replace("_stairs", "").replace("_slab", "");
-                String mcBlockName = mcBlock.getRegistryName().getPath();
+                String blockName = loc.replace("_wall", "").replace("_stairs", "").replace("_slab", "");
+                String mcBlockName = mcBlock.getDescriptionId().replace("minecraft:", "");
 
-                if (mcBlockName.equals(blockName))  {
+                if (mcBlockName.equals(blockName)) {
                     vanillaBLock = mcBlock;
+                    String vanillaBlockName = vanillaBLock.getDescriptionId().replace("blocks.minecraft.", "");
                     if (!loc.contains("wool")) {
                         if (loc.contains("stairs")) {
-                            stonecutterResultFromBase(consumer, block.get(), vanillaBLock);
+                            SingleItemRecipeBuilder.stonecutting(Ingredient.of(vanillaBLock), block.get()).unlockedBy("has_item", has(vanillaBLock)).save(consumer, loc + "_from_" + vanillaBlockName + "_stonecutting");
                             ShapedRecipeBuilder.shaped(block.get(), 4) // result
                                     .define('#', vanillaBLock) // ingredient
                                     .pattern("#  ")
@@ -43,17 +46,15 @@ public class GenRecipeProvider extends RecipeProvider {
                                     .pattern("###")
                                     .unlockedBy("has_item", has(vanillaBLock.asItem()))
                                     .save(consumer, new ResourceLocation(AdditionalVanillaStuff.MOD_ID, "shaped_" + loc));
-                        }
-                        else if (loc.contains("slab")) {
-                            stonecutterResultFromBase(consumer, block.get(), vanillaBLock, 2);
+                        } else if (loc.contains("slab")) {
+                            SingleItemRecipeBuilder.stonecutting(Ingredient.of(vanillaBLock), block.get(), 2).unlockedBy("has_item", has(vanillaBLock)).save(consumer, loc + "_from_" + vanillaBlockName + "_stonecutting");
                             ShapedRecipeBuilder.shaped(block.get(), 6) // result
                                     .define('#', vanillaBLock) // ingredient
                                     .pattern("###")
                                     .unlockedBy("has_item", has(vanillaBLock.asItem()))
                                     .save(consumer, new ResourceLocation(AdditionalVanillaStuff.MOD_ID, "shaped_" + loc));
-                        }
-                        else if (loc.contains("wall")) {
-                            stonecutterResultFromBase(consumer, block.get(), vanillaBLock);
+                        } else if (loc.contains("wall")) {
+                            SingleItemRecipeBuilder.stonecutting(Ingredient.of(vanillaBLock), block.get()).unlockedBy("has_item", has(vanillaBLock)).save(consumer, loc + "_from_" + vanillaBlockName + "_stonecutting");
                             ShapedRecipeBuilder.shaped(block.get(), 6) // result
                                     .define('#', vanillaBLock) // ingredient
                                     .pattern("###")
@@ -61,8 +62,7 @@ public class GenRecipeProvider extends RecipeProvider {
                                     .unlockedBy("has_item", has(vanillaBLock.asItem()))
                                     .save(consumer, new ResourceLocation(AdditionalVanillaStuff.MOD_ID, "shaped_" + loc));
                         }
-                    }
-                    else if (loc.contains("wool")) {
+                    } else if (loc.contains("wool")) {
                         if (loc.contains("stairs")) {
                             ShapedRecipeBuilder.shaped(block.get(), 4) // result
                                     .define('#', vanillaBLock) // ingredient
@@ -89,16 +89,5 @@ public class GenRecipeProvider extends RecipeProvider {
                 }
             }
         }
-
-/*
-                    ShapedRecipeBuilder.shaped(blocks.get(i).get()) // result
-                            .define('#', blocks.get(i - 1).get()) // ingredient
-                            .pattern("###")
-                            .pattern("###")
-                            .pattern("###")
-                            .unlockedBy("has_item", has(blocks.get(i - 1).get().asItem()))
-                            .save(consumer, new ResourceLocation("compressedblocks", "shaped_" + blockName.replace("compressedblocks:", "")));
-/
     }
 }
-*/